+++ job-server/src/main/scala/spark/jobserver/routes/DataRoutes.scala
@@ -5,7 +5,6 @@ import akka.actor.ActorRef
 import akka.pattern.ask
 import akka.util.Timeout
 import spark.jobserver.DataManagerActor._
-
 import akka.http.scaladsl.server.Route
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport.sprayJsonMarshaller
 import akka.http.scaladsl.model.StatusCodes
@@ -13,8 +12,9 @@ import spray.json.DefaultJsonProtocol._
 
 import scala.concurrent.ExecutionContext
 import scala.util.{Failure, Success}
-
 import akka.http.scaladsl.server.Directives._
+import spark.jobserver.WebApiUtils
+import spark.jobserver.WebApiUtils.successMap
 import spark.jobserver.common.akka.web.JsonUtils._
 
 /**
@@ -40,7 +40,7 @@ trait DataRoutes {
     get {
       onComplete((dataManager ? ListData).mapTo[collection.Set[String]]){
         case Success(names) => complete(names)
-        case Failure(ex) => complete(StatusCodes.InternalServerError, errMap(ex, "ERROR"))
+        case Failure(ex) => complete(StatusCodes.InternalServerError, WebApiUtils.errMap(ex, "ERROR"))
       }
     } ~
       // DELETE /data/filename delete the given file
@@ -51,10 +51,10 @@ trait DataRoutes {
             case Deleted =>
               complete(StatusCodes.OK)
             case Error =>
-              complete(StatusCodes.BadRequest, errMap(s"Unable to delete data file '$filename'."))
+              complete(StatusCodes.BadRequest, WebApiUtils.errMap(s"Unable to delete data file '$filename'."))
           }
           case Failure(ex) =>
-            complete(StatusCodes.InternalServerError, errMap(ex, "ERROR"))
+            complete(StatusCodes.InternalServerError, WebApiUtils.errMap(ex, "ERROR"))
         }
         }
       } ~
@@ -73,10 +73,10 @@ trait DataRoutes {
                         complete(StatusCodes.OK, successMap("Data reset"))
 
                       case Error =>
-                        complete(StatusCodes.BadRequest, errMap("Unable to delete data folder"))
+                        complete(StatusCodes.BadRequest, WebApiUtils.errMap("Unable to delete data folder"))
                     }
                     case Failure(ex) =>
-                      complete(StatusCodes.InternalServerError, errMap(ex, "ERROR"))
+                      complete(StatusCodes.InternalServerError, WebApiUtils.errMap(ex, "ERROR"))
                   }
                 }
               case _ => complete("ERROR")
@@ -94,10 +94,11 @@ trait DataRoutes {
                 val map = Map[String, Any](ResultKey -> Map("filename" -> name))
                 complete(StatusCodes.OK, map)
               case Error =>
-                complete(StatusCodes.BadRequest, errMap(s"Failed to store data file '$filename'."))
+                complete(StatusCodes.BadRequest,
+                  WebApiUtils.errMap(s"Failed to store data file '$filename'."))
             }
             case Failure(ex) =>
-              complete(StatusCodes.InternalServerError, errMap(ex, "ERROR"))
+              complete(StatusCodes.InternalServerError, WebApiUtils.errMap(ex, "ERROR"))
           }
 
           }
